refactor #6 : review 페이지 별점 기능 컴포넌트화
